+++ inc/template-tags.php
@@ -2,7 +2,15 @@
 /**
  * Template tags and helpers
  *
+ * File for template related things.
+ *
+ * @Author:		Roni Laukkarinen
+ * @Date:   		2021-05-18 10:29:32
+ * @Last Modified by:   Roni Laukkarinen
+ * @Last Modified time: 2022-02-08 17:18:22
+ *
  * @package air-light
+ * @link https://developer.wordpress.org/themes/basics/template-files/#template-partials
  */
 
 namespace Air_Light;
@@ -11,3 +19,4 @@ require get_theme_file_path( '/inc/template-tags/acf-blocks.php' );
 require get_theme_file_path( '/inc/template-tags/entry-footer.php' );
 require get_theme_file_path( '/inc/template-tags/single-comment.php' );
 require get_theme_file_path( '/inc/template-tags/wp-body-open.php' );
+require get_theme_file_path( '/inc/template-tags/edit-link.php' );